LayoutTests/platform/ios-sim-deprecated/media/audio-width.html

<head>
    <meta name="viewport" content="initial-scale=1">
    <script src="../../../media/media-file.js" type="text/javascript"></script>
    <script src="../../../media/video-test.js" type="text/javascript"></script>
    <style>
        #aud_1 {
            width: 44px;
        }

        #aud_2 {
            width: 400px;
        }
    </style>
    <script type="text/javascript">
        var width1 = NaN;
        var width2 = NaN;
        var player1 = null;
        var player2 = null;

        function init() {
            player1 = document.getElementById("aud_1");
            player2 = document.getElementById("aud_2");
            var audioPath = findMediaFile("audio", "../../../media/content/test");
            player1.src = audioPath;
            player2.src = audioPath;
            width1 = parseInt(window.getComputedStyle(player1).width);
            width2 = parseInt(window.getComputedStyle(player2).width);
            testExpected("width1", 260, "==");
            testExpected("width2", 400, "==");
            endTest();
        }
    </script>
</head>

<body>
    <p>Test that audio elements have a minimum width.<p>
    <audio id="aud_1" type="audio/mp4" controls></audio>
    <audio id="aud_2" type="audio/mp4" controls></audio>
    <script type="text/javascript">
        document.body.onload = init;
    </script>
</body>
